feat: refactor error handling by separating ErrorContent into its own component
